+++ BE-MyCarMaster/src/main/java/softeer/be_my_car_master/api/model/usecase/get_models/GetModelsPort.java
@@ -1,13 +1,12 @@
-package softeer.be_my_car_master.api.model.usecase.port;
+package softeer.be_my_car_master.api.model.usecase.get_models;
 
 import java.util.List;
-import java.util.Optional;
 
 import softeer.be_my_car_master.domain.model.Model;
 import softeer.be_my_car_master.global.annotation.Port;
 
 @Port
-public interface ModelPort {
+public interface GetModelsPort {
 
 	List<Model> findModels();
 }
